+++ griptape/drivers/prompt/base_prompt_driver.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Optional, Callable
+from typing import TYPE_CHECKING, Optional
 from collections.abc import Iterator
 from attrs import define, field, Factory
 from griptape.events import StartPromptEvent, FinishPromptEvent, CompletionChunkEvent
@@ -32,9 +32,6 @@ class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
     temperature: float = field(default=0.1, kw_only=True, metadata={"serializable": True})
     max_tokens: Optional[int] = field(default=None, kw_only=True, metadata={"serializable": True})
     structure: Optional[Structure] = field(default=None, kw_only=True)
-    prompt_stack_to_string: Callable[[PromptStack], str] = field(
-        default=Factory(lambda self: self.default_prompt_stack_to_string_converter, takes_self=True), kw_only=True
-    )
     ignored_exception_types: tuple[type[Exception], ...] = field(
         default=Factory(lambda: (ImportError, ValueError)), kw_only=True
     )
@@ -42,23 +39,12 @@ class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
     tokenizer: BaseTokenizer
     stream: bool = field(default=False, kw_only=True, metadata={"serializable": True})
 
-    def max_output_tokens(self, text: str | list) -> int:
-        tokens_left = self.tokenizer.count_output_tokens_left(text)
-
-        if self.max_tokens:
-            return min(self.max_tokens, tokens_left)
-        else:
-            return tokens_left
-
-    def token_count(self, prompt_stack: PromptStack) -> int:
-        return self.tokenizer.count_tokens(self.prompt_stack_to_string(prompt_stack))
-
     def before_run(self, prompt_stack: PromptStack) -> None:
         if self.structure:
             self.structure.publish_event(
                 StartPromptEvent(
                     model=self.model,
-                    token_count=self.token_count(prompt_stack),
+                    token_count=self.tokenizer.count_tokens(self.prompt_stack_to_string(prompt_stack)),
                     prompt_stack=prompt_stack,
                     prompt=self.prompt_stack_to_string(prompt_stack),
                 )
@@ -67,7 +53,9 @@ class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
     def after_run(self, result: TextArtifact) -> None:
         if self.structure:
             self.structure.publish_event(
-                FinishPromptEvent(model=self.model, token_count=result.token_count(self.tokenizer), result=result.value)
+                FinishPromptEvent(
+                    model=self.model, result=result.value, token_count=self.tokenizer.count_tokens(result.value)
+                )
             )
 
     def run(self, prompt_stack: PromptStack) -> TextArtifact:
@@ -92,7 +80,16 @@ class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
         else:
             raise Exception("prompt driver failed after all retry attempts")
 
-    def default_prompt_stack_to_string_converter(self, prompt_stack: PromptStack) -> str:
+    def prompt_stack_to_string(self, prompt_stack: PromptStack) -> str:
+        """Converts a Prompt Stack to a string for token counting or model input.
+        This base implementation is only a rough approximation, and should be overridden by subclasses with model-specific tokens.
+
+        Args:
+            prompt_stack: The Prompt Stack to convert to a string.
+
+        Returns:
+            A single string representation of the Prompt Stack.
+        """
         prompt_lines = []
 
         for i in prompt_stack.inputs: