+++ griptape/drivers/prompt/cohere_prompt_driver.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 from collections.abc import Iterator
 from attrs import define, field, Factory
 from griptape.artifacts import TextArtifact
 from griptape.drivers import BasePromptDriver
-from griptape.tokenizers import CohereTokenizer
 from griptape.utils import PromptStack, import_optional_dependency
+from griptape.tokenizers import BaseTokenizer, CohereTokenizer
 
 if TYPE_CHECKING:
     from cohere import Client
@@ -18,7 +18,6 @@ class CoherePromptDriver(BasePromptDriver):
         api_key: Cohere API key.
         model: 	Cohere model name.
         client: Custom `cohere.Client`.
-        tokenizer: Custom `CohereTokenizer`.
     """
 
     api_key: str = field(kw_only=True, metadata={"serializable": False})
@@ -27,7 +26,7 @@ class CoherePromptDriver(BasePromptDriver):
         default=Factory(lambda self: import_optional_dependency("cohere").Client(self.api_key), takes_self=True),
         kw_only=True,
     )
-    tokenizer: CohereTokenizer = field(
+    tokenizer: BaseTokenizer = field(
         default=Factory(lambda self: CohereTokenizer(model=self.model, client=self.client), takes_self=True),
         kw_only=True,
     )
@@ -44,26 +43,23 @@ class CoherePromptDriver(BasePromptDriver):
             if event.event_type == "text-generation":
                 yield TextArtifact(value=event.text)
 
+    def _prompt_stack_input_to_message(self, prompt_input: PromptStack.Input) -> dict:
+        if prompt_input.is_system():
+            return {"role": "SYSTEM", "text": prompt_input.content}
+        elif prompt_input.is_user():
+            return {"role": "USER", "text": prompt_input.content}
+        else:
+            return {"role": "ASSISTANT", "text": prompt_input.content}
+
     def _base_params(self, prompt_stack: PromptStack) -> dict:
         user_message = prompt_stack.inputs[-1].content
-        history_messages = [self.__to_cohere_message(input) for input in prompt_stack.inputs[:-1]]
+
+        history_messages = [self._prompt_stack_input_to_message(input) for input in prompt_stack.inputs[:-1]]
 
         return {
             "message": user_message,
             "chat_history": history_messages,
             "temperature": self.temperature,
             "stop_sequences": self.tokenizer.stop_sequences,
+            "max_tokens": self.max_tokens,
         }
-
-    def __to_cohere_message(self, input: PromptStack.Input) -> dict[str, Any]:
-        return {"role": self.__to_cohere_role(input.role), "text": input.content}
-
-    def __to_cohere_role(self, role: str) -> str:
-        if role == PromptStack.SYSTEM_ROLE:
-            return "SYSTEM"
-        if role == PromptStack.USER_ROLE:
-            return "USER"
-        elif role == PromptStack.ASSISTANT_ROLE:
-            return "CHATBOT"
-        else:
-            return "USER"