+++ griptape/tasks/prompt_task.py
@@ -34,7 +34,7 @@ class PromptTask(BaseTextInputTask):
 
         if memory:
             # inserting at index 1 to place memory right after system prompt
-            stack.add_conversation_memory(memory, 1)
+            memory.add_to_prompt_stack(stack, 1)
 
         return stack