+++ griptape/utils/__init__.py
@@ -14,9 +14,9 @@ from .hash import str_to_hash
 from .import_utils import import_optional_dependency
 from .import_utils import is_dependency_installed
 from .stream import Stream
-from .constants import Constants as constants
 from .load_artifact_from_memory import load_artifact_from_memory
 from .deprecation import deprecation_warn
+from .structure_visualizer import StructureVisualizer
 
 
 def minify_json(value: str) -> str:
@@ -40,9 +40,9 @@ __all__ = [
     "remove_null_values_in_dict_recursively",
     "dict_merge",
     "Stream",
-    "constants",
     "load_artifact_from_memory",
     "deprecation_warn",
     "load_file",
     "load_files",
+    "StructureVisualizer",
 ]