+++ griptape/utils/constants.py
@@ -1,4 +1,0 @@
-class Constants:
-    # Stop sequence for chain-of-thought in the framework. Using this "token-like" string to make it more unique,
-    # so that it doesn't trigger on accident.
-    RESPONSE_STOP_SEQUENCE = "<|Response|>"