+++ tests/unit/structures/test_workflow.py
@@ -1,16 +1,34 @@
+import time
 import pytest
 
+from pytest import fixture
 from griptape.memory.task.storage import TextArtifactStorage
 from tests.mocks.mock_prompt_driver import MockPromptDriver
 from griptape.rules import Rule, Ruleset
-from griptape.tasks import PromptTask, BaseTask, ToolkitTask
+from griptape.tasks import PromptTask, BaseTask, ToolkitTask, CodeExecutionTask
 from griptape.structures import Workflow
+from griptape.artifacts import ErrorArtifact, TextArtifact
 from griptape.memory.structure import ConversationMemory
 from tests.mocks.mock_tool.tool import MockTool
 from tests.mocks.mock_embedding_driver import MockEmbeddingDriver
 
 
 class TestWorkflow:
+    @fixture
+    def waiting_task(self):
+        def fn(task):
+            time.sleep(10)
+            return TextArtifact("done")
+
+        return CodeExecutionTask(run_fn=fn)
+
+    @fixture
+    def error_artifact_task(self):
+        def fn(task):
+            return ErrorArtifact("error")
+
+        return CodeExecutionTask(run_fn=fn)
+
     def test_init(self):
         driver = MockPromptDriver()
         workflow = Workflow(prompt_driver=driver, rulesets=[Ruleset("TestRuleset", [Rule("test")])])
@@ -676,6 +694,7 @@ class TestWorkflow:
         context = workflow.context(task)
 
         assert context["parent_outputs"] == {parent.id: parent.output.to_text()}
+        assert context["parents_output_text"] == "mock output"
         assert context["structure"] == workflow
         assert context["parents"] == {parent.id: parent}
         assert context["children"] == {child.id: child}
@@ -689,6 +708,14 @@ class TestWorkflow:
 
         with pytest.deprecated_call():
             Workflow(stream=True)
+
+    def test_run_with_error_artifact(self, error_artifact_task, waiting_task):
+        end_task = PromptTask("end")
+        end_task.add_parents([error_artifact_task, waiting_task])
+        workflow = Workflow(prompt_driver=MockPromptDriver(), tasks=[waiting_task, error_artifact_task, end_task])
+        workflow.run()
+
+        assert workflow.output is None
 
     @staticmethod
     def _validate_topology_1(workflow):